add -wait-and-cleanup and signal handler on all platforms
